+++ tfe/data_source_variable_set_test.go
@@ -56,6 +56,8 @@ func TestAccTFEVariableSetsDataSource_full(t *testing.T) {
 						"data.tfe_variable_set.foobar", "workspace_ids.#", "1"),
 					resource.TestCheckResourceAttr(
 						"data.tfe_variable_set.foobar", "variable_ids.#", "1"),
+					resource.TestCheckResourceAttr(
+						"data.tfe_variable_set.foobar", "project_ids.#", "1"),
 				),
 			},
 		},
@@ -94,11 +96,21 @@ resource "tfe_workspace" "foobar" {
   organization = tfe_organization.foobar.id
 }
 
+resource "tfe_project" "foobar" {
+  name         = "project-foo-%d"
+  organization = tfe_organization.foobar.id
+}
+
 resource "tfe_variable_set" "foobar" {
   name = "varset-foo-%d"
 	description = "a description"
 	organization = tfe_organization.foobar.id
 	workspace_ids = [tfe_workspace.foobar.id]
+}
+
+resource "tfe_project_variable_set" "foobar" {
+	variable_set_id = tfe_variable_set.foobar.id
+	project_id = tfe_project.foobar.id
 }
 
 resource "tfe_variable" "envfoo" {
@@ -112,5 +124,5 @@ data "tfe_variable_set" "foobar" {
   name = tfe_variable_set.foobar.name
 	organization = tfe_variable_set.foobar.organization
 	depends_on = [tfe_variable.envfoo]
-}`, rInt, rInt, rInt)
+}`, rInt, rInt, rInt, rInt)
 }