+++ CHANGELOG.md
@@ -4,6 +4,7 @@ BUG FIXES:
 * `r/tfe_team_project_access`: Fixes a panic that occurs when the client is configured against an older TFE release, by @sebasslash [1011](https://github.com/hashicorp/terraform-provider-tfe/pull/1011)
 * The provider no longer makes two service discovery requests per provider config, by @brandonc [1034](https://github.com/hashicorp/terraform-provider-tfe/pull/1034)
 * `d/tfe_policy_set`: Add `excluded_workspace_ids` attribute, by @Netra2104 [1035](https://github.com/hashicorp/terraform-provider-tfe/pull/1035)
+* `r/tfe_workspace_policy_set` and `r/tfe_workspace_policy_set_exclusion`: Removed workspace-related `include` parameters since included workspace data is not required to manage these resources, by @jbonhag [1080](https://github.com/hashicorp/terraform-provider-tfe/pull/1080)
 
 FEATURES:
 * `d/tfe_organization_membership`: Add `organization_membership_id` attribute, by @laurenolivia [997](https://github.com/hashicorp/terraform-provider-tfe/pull/997)